refactor(schema-generator): keep one copy of the UI-contract hint helpers
Two files carried the same pair of functions, character for character apart
from brace placement. `schema-generator.ts` had a private `attachUiContract`
and did the hint lookup inline in `applyNodeSchemaHints`;
`formatters/object-formatter.ts` had a module-level `attachUiContract` plus a
`getUiContractHint` that performed the same lookup, including the same fallback
to `ts.getOriginalNode`. Both spell out the UI-contract shape inline, so a
change to that shape had to be made twice, in step, or the two would drift.

They move to a new `src/ui-contract.ts` and both callers import them. The
lookup helper already defaulted its node argument to `context.typeNode`, which
is exactly what the generator's inline version used, so `applyNodeSchemaHints`
becomes a two-line call.

No behavior change: the surviving copy is the object formatter's, and the
generator's copy differed from it only in formatting.

Verified with `deno task check`, `deno task test` and `deno lint` in
`packages/schema-generator`.
